feat: Allow attacks to run without an RPC endpoint (#107)
* rpc-less workspace test

* add: block caches

* add: foundry cache

* edit: attack block numbers

* add: node support for devcontainers

* add: cache warming and rpc serving

* edit: readme to include warming instuctions

* feat: remove unnecessary dependency of underscored block numbers

* chore: improve cache dirs and org

* feat: support for new dir structure and underscored block numbers

* edit: readme

* feat: expand cache

* feat: allow re-running attacks with custom RPC endpoint

* edit: readme

* feat: improve rpc fallback logs

* chore: cleanup test files

+++ test/Business_Logic/FourMeme/IFourMemeToken.sol
@@ -1,7 +1,6 @@
-import {IERC20} from '../../interfaces/IERC20.sol';
+import {IERC20} from "../../interfaces/IERC20.sol";
 pragma solidity ^0.8.17;
 
 interface IFourMemeToken is IERC20 {
-
-    function setMode(uint v) external;
+    function setMode(uint256 v) external;
 }
