+++ test/utils/IPancakeRouter01.sol
@@ -2,8 +2,11 @@
 pragma solidity ^0.8.17;
 
 interface IPancakeRouter01 {
-    
-    function swapExactTokensForETH(uint amountIn, uint amountOutMin, address[] memory path, address to, uint deadline)
-        external
-        returns (uint[] memory amounts);
+    function swapExactTokensForETH(
+        uint256 amountIn,
+        uint256 amountOutMin,
+        address[] memory path,
+        address to,
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 }
